+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/FrameworkExtensionTestCase.php
@@ -176,6 +176,7 @@ abstract class FrameworkExtensionTestCase extends TestCase
         $def = $container->getDefinition('property_accessor');
         $this->assertSame(PropertyAccessor::MAGIC_SET | PropertyAccessor::MAGIC_GET, $def->getArgument(0));
         $this->assertSame(PropertyAccessor::THROW_ON_INVALID_PROPERTY_PATH, $def->getArgument(1));
+        $this->assertFalse($def->getArgument(5));
     }
 
     public function testPropertyAccessWithOverriddenValues()
@@ -184,6 +185,7 @@ abstract class FrameworkExtensionTestCase extends TestCase
         $def = $container->getDefinition('property_accessor');
         $this->assertSame(PropertyAccessor::MAGIC_GET | PropertyAccessor::MAGIC_CALL, $def->getArgument(0));
         $this->assertSame(PropertyAccessor::THROW_ON_INVALID_INDEX, $def->getArgument(1));
+        $this->assertTrue($def->getArgument(5));
     }
 
     public function testPropertyAccessCache()