+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -42,6 +42,7 @@ CHANGELOG
  * Add support for rate limited transports
  * Register the security functions `is_granted()`, `is_authenticated()`, `is_fully_authenticated()`, `is_remember_me()` and `current_user()` in the validator expression language
  * Add `framework.profiler.excluded_paths` and `framework.profiler.excluded_http_codes` to skip profiling requests matching a path or answered with a given HTTP status code
+ * Add `framework.property_access.wildcard_reads` option to read every element of a collection through a `[*]` path
 
 8.1
 ---